backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp==3.9.3",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asyncer==0.0.8",
    "attrs==25.3.0",
    "autogen-core==0.5.5",
    "banks==2.1.2",
    "beautifulsoup4==4.13.4",
    "black==24.1.1",
    "bs4==0.0.2",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "dataclasses-json==0.6.7",
    "deprecated==1.2.18",
    "dirtyjson==1.0.8",
    "diskcache==5.6.3",
    "distro==1.9.0",
    "docker==7.1.0",
    "fastapi==0.109.2",
    "filelock==3.18.0",
    "filetype==1.2.0",
    "flake8==7.0.0",
    "frozenlist==1.6.0",
    "fsspec==2025.3.2",
    "google-genai>=0.7.0",
    "googleapis-common-protos==1.70.0",
    "greenlet==3.2.1",
    "griffe==1.7.3",
    "grpcio==1.71.0",
    "grpcio-status==1.62.3",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "huggingface-hub==0.30.2",
    "idna==3.10",
    "importlib-metadata==8.6.1",
    "iniconfig==2.1.0",
    "isort==5.13.2",
    "jinja2==3.1.6",
    "jiter==0.9.0",
    "joblib==1.5.0",
    "jsonref==1.1.0",
    "llama-index-core==0.12.34.post1",
    "llama-index-embeddings-huggingface==0.5.3",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "mccabe==0.7.0",
    "mpmath==1.3.0",
    "multidict==6.4.3",
    "mypy-extensions==1.1.0",
    "nest-asyncio==1.6.0",
    "networkx==3.4.2",
    "nltk==3.9.1",
    "numpy==2.2.5",
    "openai==1.76.2",
    "opentelemetry-api==1.32.1",
    "packaging==25.0",
    "pathspec==0.12.1",
    "pillow==11.2.1",
    "platformdirs==4.3.7",
    "pluggy==1.5.0",
    "propcache==0.3.1",
    "proto-plus==1.26.1",
    "protobuf==5.29.4",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pyautogen==0.9",
    "pycodestyle==2.11.1",
    "pydantic==2.11.3",
    "pydantic-core==2.33.1",
    "pydantic-settings==2.9.1",
    "pyflakes==3.2.0",
    "pyparsing==3.2.3",
    "pytest==7.4.4",
    "pytest-asyncio==0.23.5",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.7",
    "pywin32==310",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "rsa==4.9.1",
    "safetensors==0.5.3",
    "scikit-learn==1.6.1",
    "scipy==1.15.2",
    "sentence-transformers==4.1.0",
    "setuptools==80.3.0",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "sqlalchemy==2.0.40",
    "starlette==0.36.3",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "termcolor==3.1.0",
    "threadpoolctl==3.6.0",
    "tiktoken==0.9.0",
    "tokenizers==0.21.1",
    "torch==2.7.0",
    "tqdm==4.67.1",
    "transformers==4.51.3",
    "typing-extensions==4.13.2",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.0",
    "ujson==5.9.0",
    "uritemplate==4.1.1",
    "urllib3==2.4.0",
    "uvicorn==0.27.1",
    "wrapt==1.17.2",
    "yarl==1.20.0",
    "zipp==3.21.0",
]
